requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
data==0.4
decorator==5.1.1
Deprecated==1.2.13
Django==4.1.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.3.1
funcsigs==1.0.2
h11==0.12.0
httpcore==0.15.0
idna==3.4
kombu==5.2.4
multidict==6.0.2
numpy==1.23.3
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
yarl==1.8.1
